fix(frontend): move Protection section to bottom, fix save bar visibility (#232)
- Extract ProtectionSection (lock/delete) from SettingsTabPanel into
  standalone component rendered after SearchSettingsSection
- Fix save bar always visible: render conditionally (isDirty &&) instead
  of CSS translate trick which fails for non-zero bottom offset
